Fix scrub aborts with No such process (issue 19)
Used an int instead of an off_t to hold file size.
This was introduced with fillfile threading in scrub-2.5.0.
Failure is immediate and obvious for targets > 2G.

+++ src/fillfile.c
@@ -74,7 +74,7 @@ refill_thread(void *arg)
 
 static int
 refill_memcpy(struct memstruct *mp, unsigned char *mem, int memsize,
-              int filesize, int written)
+              off_t filesize, off_t written)
 {
 #if WITH_PTHREADS
     if (no_threads) {
